bugs fixed

+++ client/src/app/app.component.ts
@@ -74,7 +74,8 @@ export class AppComponent implements OnInit {
 
   deleteExpense($event: any) {
     this.expenseService.deleteExpense($event).subscribe((response) => {
-      this.expenseService.deleteFile(response.data.billFile).subscribe();
+      if(response.data.bill_file)
+        this.expenseService.deleteFile(response.data.billFile).subscribe();
       this.expenses = this.expenses.filter(exp => {
         return exp._id !== $event;
       });
