+++ client/src/app/components/form/form.component.ts
@@ -23,6 +23,7 @@ export class FormComponent implements OnChanges, OnInit {
   @ViewChild('fileInput') fileInput!: ElementRef;
   isFileSelected: boolean = false;
   isFileUploaded: boolean = false;
+  isFileDeleted: boolean = false;
   billFile!: File;
   uploadedFileName: string = "";
 
@@ -104,7 +105,7 @@ export class FormComponent implements OnChanges, OnInit {
       newExpense._id = this.formOptions.editExpenseId;
       this.onEditExpense.emit(newExpense);
     }
-
+    this.resetForm();
     this.closeForm(null);
   }
 
@@ -144,23 +145,32 @@ export class FormComponent implements OnChanges, OnInit {
     });
   }
   fileDelete() {
-    this.expenseService.deleteFile(this.uploadedFileName).subscribe(response => {
-      this.uploadedFileName = "";
-      this.isFileSelected = false;
-      this.isFileUploaded = false;
-    });
+    if (confirm("Do you want to delete this file?")) {
+      this.expenseService.deleteFile(this.uploadedFileName).subscribe(response => {
+        this.uploadedFileName = "";
+        this.isFileSelected = false;
+        this.isFileUploaded = false;
+        this.isFileDeleted = true;
+      });
+    }
   }
 
   resetForm() {
     this.isFileUploaded = false
     this.isFileSelected = false;
+    this.isFileDeleted = false;
     this.expenseForm.reset();
   }
 
   closeForm(event:any) {
     if(event && event.target != event.currentTarget)
-      return
+      return;
     
+    if(this.isFileUploaded || this.isFileDeleted){
+      alert("Please save this expense");
+      return;
+    }
+
     this.resetForm();
 
     this.getFormStatus.emit({
